@@ -21,15 +21,14 @@
         "issues": "https://github.com/Blog404DE/RegenRadarVideo/issues"
     },
     "require-dev": {
-        "roave/security-advisories": "dev-master",
-        "phpmd/phpmd": "2.8.*",
-        "sebastian/phpcpd": "4.1.*",
-        "squizlabs/php_codesniffer": "3.5.*",
-        "friendsofphp/php-cs-fixer": "2.17.*",
-        "wapmorgan/php-deprecation-detector": "^2.0"
+        "roave/security-advisories": "dev-latest",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "sebastian/phpcpd": "^6.0",
+        "wapmorgan/php-deprecation-detector": "dev-master",
+        "squizlabs/php_codesniffer": "^3.6"
     },
     "require": {
-        "php": "^7.1 <=7.5",
+        "php": "^8.0",
         "ext-curl": "*",
         "ext-mbstring": "*"
     },
