@@ -1,5 +1,6 @@
 #!/usr/bin/env php
 <?php
+
 /*
  *  RegenRadar Generator aus DWD-Radarbilder from neuthardwetter.de by Jens Dutzi
  *
@@ -10,6 +11,7 @@
  *  @version    3.3.1
  *  @link       https://github.com/Blog404DE/RegenRadarVideo
  */
+
 use blog404de\RegenRadar\RegenRadar;
 
 try {
@@ -40,19 +42,19 @@ try {
     // Erzeuge Radar-Videos
     foreach ($config as $value) {
         $header = 'Beginne mit dem erzeugen des Radar-Video vom DWD Webserver: ' . basename($value['remoteURL']);
-        echo PHP_EOL . $header . PHP_EOL . str_repeat('=', \mb_strlen($header)) . PHP_EOL;
+        echo PHP_EOL . $header . PHP_EOL . str_repeat('=', mb_strlen($header)) . PHP_EOL;
 
         // Lokaler Dateiname für das Download
         $localVideoFile = $value['localFolder'] . '/' . basename($value['remoteURL']);
 
-        // Prüfe DWD Video nach potentiellen Updates
+        // Prüfe DWD Video nach potenziellen Updates
         echo 'Prüfe ob Update des Videos notwendig ist:' . PHP_EOL;
         if (file_exists($localVideoFile)) {
-            if (!$value['forceRebuild']) {
-                $needRebuild = $regenradarBot->network->checkDWDRadarVideoForUpdate($localVideoFile, $value['remoteURL']);
-            } else {
+            if ($value['forceRebuild']) {
                 echo '-> Update des Videos ist wurde erzwungen durch die Konfigurationsdatei' . PHP_EOL;
                 $needRebuild = true;
+            } else {
+                $needRebuild = $regenradarBot->network->checkDWDRadarVideoForUpdate($localVideoFile, $value['remoteURL']);
             }
         } else {
             echo '-> Update des Videos ist notwendig, da noch keine lokale Datei existiert' . PHP_EOL;
@@ -73,7 +75,7 @@ try {
             foreach ($value['output'] as $filetype => $filename) {
                 if ('poster' !== $filetype) {
                     $header = 'Erzeuge ' . $filetype . ' aus den Radar-Daten';
-                    echo PHP_EOL . $header . PHP_EOL . str_repeat('=', \mb_strlen($header)) . PHP_EOL;
+                    echo PHP_EOL . $header . PHP_EOL . str_repeat('=', mb_strlen($header)) . PHP_EOL;
                     if (empty($filename) || false === $filename) {
                         echo '-> Erzeugen des ' . $filetype . '-Videos in der Konfiguration deaktiviert' . PHP_EOL;
                     } else {
@@ -82,12 +84,12 @@ try {
                     }
                 } else {
                     $header = 'Lade Poster-Datei für das Video vom DWD Webserver';
-                    echo PHP_EOL . $header . PHP_EOL . str_repeat('=', \mb_strlen($header)) . PHP_EOL;
+                    echo PHP_EOL . $header . PHP_EOL . str_repeat('=', mb_strlen($header)) . PHP_EOL;
 
-                    if (false === $value['posterURL']) {
-                        echo PHP_EOL . '-> Download der Poster-Datei wird nicht benötigt' . PHP_EOL;
-                    } else {
+                    if (false !== $value['posterURL']) {
                         $regenradarBot->network->downloadFile($filename, $value['posterURL']);
+                    } else {
+                        echo PHP_EOL . '-> Download der Poster-Datei wird nicht benötigt' . PHP_EOL;
                     }
                 }
             }
@@ -95,7 +97,7 @@ try {
 
         echo PHP_EOL . '... Auftrag ausgeführt!' . PHP_EOL . PHP_EOL;
     }
-} catch (RuntimeException | \Exception $e) {
+} catch (RuntimeException|\Exception $e) {
     // Fehler-Handling
     fwrite(STDERR, 'Fataler Fehler: ' . $e->getFile() . ':' . $e->getLine() . ' - ' . $e->getMessage());
 
